requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
Werkzeug==2.2.3
zipp==3.15.0
